refactor: extract option/localstorage
